v.0.1 Перевод приложения на androidX Исправление бага с доступом к изображениям на Android Q Новая верстка экрана Случайная карта

+++ app/src/main/java/ru/mgusev/eldritchhorror/repository/FileHelper.java
@@ -1,18 +1,21 @@
 package ru.mgusev.eldritchhorror.repository;
 
+import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.net.Uri;
 import android.os.Environment;
 import android.os.Parcelable;
+import android.provider.MediaStore;
 
 import com.facebook.common.util.UriUtil;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.channels.FileChannel;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -28,6 +31,25 @@ public class FileHelper {
         this.context = context;
     }
 
+    public static Uri getImageContentUri(Context context, File imageFile) {
+        String filePath = imageFile.getAbsolutePath();
+        Cursor cursor = context.getContentResolver().query(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, new String[] { MediaStore.Images.Media._ID },
+                MediaStore.Images.Media.DATA + "=? ", new String[] { filePath }, null);
+        if (cursor != null && cursor.moveToFirst()) {
+            int id = cursor.getInt(cursor.getColumnIndex(MediaStore.MediaColumns._ID));
+            cursor.close();
+            return Uri.withAppendedPath(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, "" + id);
+        } else {
+            if (imageFile.exists()) {
+                ContentValues values = new ContentValues();
+                values.put(MediaStore.Images.Media.DATA, filePath);
+                return context.getContentResolver().insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, values);
+            } else {
+                return null;
+            }
+        }
+    }
+
     public List<String> getImages(long gameId) {
         uriFileList = new ArrayList<>();
         File dir = context.getExternalFilesDir(Environment.DIRECTORY_PICTURES + File.separator + gameId);
@@ -37,7 +59,7 @@ public class FileHelper {
                 uriFileList.add(UriUtil.getUriForFile(file).toString());
             }
         }
-        if (uriFileList.size() == 0)
+        if (uriFileList.size() == 0 && dir != null)
             Timber.tag("DELETE FILE DIR").d(Objects.requireNonNull(dir).getPath() + " " + dir.delete());
         return uriFileList;
     }
@@ -48,11 +70,19 @@ public class FileHelper {
     }
 
     public void copyFile(File source, File dest) throws IOException {
-        FileChannel sourceChannel = new FileInputStream(source).getChannel();
-        FileChannel destChannel = new FileOutputStream(dest).getChannel();
-        destChannel.transferFrom(sourceChannel, 0, sourceChannel.size());
-        sourceChannel.close();
-        destChannel.close();
+        InputStream inputStream = context.getContentResolver().openInputStream(Objects.requireNonNull(getImageContentUri(context, source)));
+        FileOutputStream fileOutputStream = new FileOutputStream(dest);
+        copyStream(Objects.requireNonNull(inputStream), fileOutputStream);
+        fileOutputStream.close();
+        inputStream.close();
+    }
+
+    private static void copyStream(InputStream input, OutputStream output) throws IOException {
+        byte[] buffer = new byte[1024];
+        int bytesRead;
+        while ((bytesRead = input.read(buffer)) != -1) {
+            output.write(buffer, 0, bytesRead);
+        }
     }
 
     public void deleteRecursiveFiles(File fileOrDirectory) {
